requirements.txt
requests==2.31.0
pytest==7.4.0
pytest-cov==4.1.0
python-dotenv==1.0.0
aiohttp==3.8.5
xmltodict==0.13.0
